backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
httpx
